move ordered_model to common
