+++ engine/apps/alerts/models/channel_filter.py
@@ -9,7 +9,7 @@ from django.db import models
 
 from common.jinja_templater import apply_jinja_template
 from common.jinja_templater.apply_jinja_template import JinjaTemplateError, JinjaTemplateWarning
-from common.ordered_model import OrderedModel
+from common.ordered_model.ordered_model import OrderedModel
 from common.public_primary_keys import generate_public_primary_key, increase_public_primary_key_length
 
 logger = logging.getLogger(__name__)
