+++ engine/apps/api/views/channel_filter.py
@@ -11,7 +11,6 @@ from apps.api.serializers.channel_filter import (
     ChannelFilterSerializer,
     ChannelFilterUpdateSerializer,
 )
-from apps.api.views.ordered_model import OrderedModelViewSet
 from apps.auth_token.auth import PluginAuthentication
 from apps.slack.models import SlackChannel
 from common.api_helpers.exceptions import BadRequest
@@ -22,6 +21,7 @@ from common.api_helpers.mixins import (
     UpdateSerializerMixin,
 )
 from common.insight_log import EntityEvent, write_resource_insight_log
+from common.ordered_model.viewset import OrderedModelViewSet
 
 
 class ChannelFilterView(
