+++ engine/apps/base/tests/test_ordered_model.py
@@ -4,7 +4,7 @@ import threading
 import pytest
 from django.db import models
 
-from common.ordered_model import OrderedModel
+from common.ordered_model.ordered_model import OrderedModel
 
 
 class TestOrderedModel(OrderedModel):
