+++ engine/apps/public_api/serializers/escalation_policies.py
@@ -5,7 +5,6 @@ from django.utils.functional import cached_property
 from rest_framework import fields, serializers
 
 from apps.alerts.models import CustomButton, EscalationChain, EscalationPolicy
-from apps.public_api.serializers.ordered_model import OrderedModelSerializer
 from apps.schedules.models import OnCallSchedule
 from apps.slack.models import SlackUserGroup
 from apps.user_management.models import User
@@ -16,6 +15,7 @@ from common.api_helpers.custom_fields import (
 )
 from common.api_helpers.exceptions import BadRequest
 from common.api_helpers.mixins import EagerLoadingMixin
+from common.ordered_model.serializer import OrderedModelSerializer
 
 
 class EscalationPolicyTypeField(fields.CharField):
