+++ engine/apps/public_api/serializers/routes.py
@@ -4,10 +4,10 @@ from rest_framework import fields, serializers
 from apps.alerts.models import AlertReceiveChannel, ChannelFilter, EscalationChain
 from apps.api.serializers.alert_receive_channel import valid_jinja_template_for_serializer_method_field
 from apps.base.messaging import get_messaging_backend_from_id, get_messaging_backends
-from apps.public_api.serializers.ordered_model import OrderedModelSerializer
 from common.api_helpers.custom_fields import OrganizationFilteredPrimaryKeyRelatedField
 from common.api_helpers.exceptions import BadRequest
 from common.jinja_templater.apply_jinja_template import JinjaTemplateError
+from common.ordered_model.serializer import OrderedModelSerializer
 from common.utils import is_regex_valid
